# swagger-docs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/professor": {
      "post": {
        "operationId": "ProfessorController_createProfessor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfessorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Professor created successfully"
          },
          "409": {
            "description": "Email already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new professor",
        "tags": ["Professor"]
      },
      "get": {
        "operationId": "ProfessorController_getAllProfessors",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (starts from 1)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by (name = first_name + last_name)",
            "schema": {
              "enum": ["name", "email", "created_at", "updated_at"],
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order (asc or desc)",
            "schema": {
              "enum": ["asc", "desc"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Professors retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all professors (School Admin only)",
        "tags": ["Professor"]
      }
    },
    "/api/professor/{id}": {
      "patch": {
        "operationId": "ProfessorController_updateProfessor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Professor ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfessorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Professor updated successfully"
          },
          "404": {
            "description": "Professor not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update professor details",
        "tags": ["Professor"]
      },
      "get": {
        "operationId": "ProfessorController_getProfessorById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Professor ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Professor retrieved successfully"
          },
          "404": {
            "description": "Professor not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get professor by ID (School Admin only)",
        "tags": ["Professor"]
      },
      "delete": {
        "operationId": "ProfessorController_deleteProfessor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Professor ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Professor deleted successfully"
          },
          "404": {
            "description": "Professor not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft delete a professor",
        "tags": ["Professor"]
      }
    },
    "/api/professor/password/{id}": {
      "patch": {
        "operationId": "ProfessorController_updateProfessorPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Professor ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfessorPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated successfully"
          },
          "404": {
            "description": "Professor not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update professor password",
        "tags": ["Professor"]
      }
    },
    "/api/professor/{id}/assignments": {
      "get": {
        "operationId": "ProfessorController_getProfessorAssignments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Professor ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (starts from 1)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Professor assignments retrieved successfully"
          },
          "404": {
            "description": "Professor not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get professor assignments (School Admin and Professor)",
        "tags": ["Professor"]
      }
    },
    "/api/professor/{id}/module-access/{moduleId}": {
      "get": {
        "operationId": "ProfessorController_checkProfessorModuleAccess",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Professor ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          },
          {
            "name": "moduleId",
            "required": true,
            "in": "path",
            "description": "Module ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Module access check completed"
          },
          "404": {
            "description": "Professor or module not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Check professor access to a specific module",
        "tags": ["Professor"]
      }
    },
    "/api/professor/{id}/status": {
      "patch": {
        "operationId": "ProfessorController_updateProfessorStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Professor ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Professor status updated successfully"
          },
          "404": {
            "description": "Professor not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update professor status (School Admin only)",
        "tags": ["Professor"]
      }
    },
...
  },
  "info": {
    "title": "AI Professor API",
    "description": "API documentation for AI Professor backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Status of the entity",
            "enum": ["ACTIVE", "INACTIVE"],
            "example": "ACTIVE"
          }
        },
        "required": ["status"]
      },
...
      "CreateProfessorDto": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string",
            "example": "John",
            "description": "First name of the professor"
          },
          "last_name": {
            "type": "string",
            "example": "Doe",
            "description": "Last name of the professor"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "Email address"
          },
          "school_id": {
            "type": "string",
            "example": "60d21b4667d0d8992e610c85",
            "description": "School ID (MongoDB ObjectId)"
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "INACTIVE"],
            "example": "ACTIVE",
            "description": "Professor status (ACTIVE or INACTIVE)"
          }
        },
        "required": ["first_name", "email", "school_id"]
      },
      "UpdateProfessorDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "number",
            "example": 1234567890
          },
          "country_code": {
            "type": "number",
            "example": 1
          },
          "first_name": {
            "type": "string",
            "example": "John"
          },
          "last_name": {
            "type": "string",
            "example": "Doe"
          },
          "profile_pic": {
            "type": "string",
            "example": "https://bucket.s3.amazonaws.com/profile-pics/123-uuid-filename.jpg",
            "description": "Profile picture URL from authorized upload"
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "INACTIVE"],
            "example": "ACTIVE",
            "description": "Professor status (ACTIVE or INACTIVE)"
          }
        }
      },
      "UpdateProfessorPasswordDto": {
        "type": "object",
        "properties": {
          "old_password": {
            "type": "string",
            "example": "OldPassword123!"
          },
          "new_password": {
            "type": "string",
            "example": "NewPassword123!",
            "description": "Password must be at least 8 characters with uppercase, lowercase, number, and special character"
          }
        },
        "required": ["old_password", "new_password"]
      },
...
    }
  }
}